Plugin Cleanup - remove old ide plugins and update others
Antrun to 1.8
Maven-bundle-plugin to 2.5.0
Maven-project-info-reports-plugin to 2.8

Remove maven-idea-plugin and maven-eclipse-plugin.  Both of these are
rather legacy and modern IDE's no longer need these.  In this case both
were attempting to setup download sources and javadocs.  Technically
that doesn't belong with mybatis anyways as that is a configuration
decision to be made in the IDE as to maven downloading those pieces.
However, running these plugins is likely to cause other issues as it
then also causes legacy things to occur such as messing up classpath.