Jacoco report goal is incorrectly configured
"report" is a reporting goal, not a build goal.  If necessary,
configurations to "report" should be made under the reporting section of
the pom.  There's no need to add it in under reporting in the parent
because it is executed automatically by the site goal.

Having the goal configured here causes maven to build the jacoco site
during the build which can cause problems for downstream projects that
might want to change how the jacoco site is built.