[pom] Move m2e compiler setup args to for java as a whole over when in eclipse
Tests can be higher jdk.  M2E cannot handle this type of scenario, the provided help to force up addressed source/target only on source end is insufficient.  Move this to properties in the profile to force it across the board, this will ensure source/target/release entirely and address any other plugins that need this due to usage within Eclipse.

There is always a catch-22 here.  The catch is that developers need to know the difference as main build will fail if attempting to use higher java code where not yet appropriate.  Running within Eclipse will appear ok but from command line where it matters, this will not occur and builds would fail if incorrectly used.  If M2E learns to treat src/test separately this could be removed.

This change overall should be status quo.  Since jdk in Eclipse modified in such a way, it would affect other plugins if activations for M2E suggested to do so already.  This method is just cleaner and makes it a bit more obvious.