[pom] Add hack to fix maven defect with site distributions using scm process
stating is supposed to lay files to target/staging but instead it takes the git name such as parent.git and puts that one directory up.  This hack fixes bulk of mybatis to just behave as we expect.  For the few multi module builds we have, those need to wipe this value out such that it is defined as ```<topSiteURL />```