[pom] Remove test of build support prior to java 17.
As plugins are moving to java 17, its not worth trying to support legacy especially when some like formatter plugin were corrupt for a few releases so backwards support is more difficult.  For now, we will be targeting java 11 in the parent but intend to release as java 8 one more time with all downstream plugins set to java 8.  After that those can move up to 11 then later we will move parent to 17 and continue moving up elsewhere.