Remove developer list
This data can be primarily retrieved from git history which mybatis has been in for significant number of years.  Its further a maintenance concern as large number of developers here are no longer working with the project and uncertain if data is accurate any longer.  In order to make this more support friendly, this data should generally be removed in light of commit data providing enough information on its own.  As this commit will also remain in history, data is not actually lost by makign this action.  This is similar to removing the author tags from files.